Raytracing Cuda + OMP donctionnel

+++ Student_Cuda_Image/src/cpp/core/01_Rippling/01_animable/device/math/RipplingMath.h
@@ -21,15 +21,14 @@ class RipplingMath
 	 \*-------------------------------------*/
 
     public:
-	__device__
-	RipplingMath(uint w)
+	__device__ RipplingMath(uint w)
 	    {
 	    this->dim2 = w / 2.f;
 	    }
 
 	// constructeur copie: pas besoin car pas attribut ptr
 	__device__
-	virtual ~RipplingMath(void)
+	 virtual ~RipplingMath(void)
 	    {
 	    // rien
 	    }
@@ -40,7 +39,7 @@ class RipplingMath
 
     public:
 	__device__
-	void colorIJ(uchar4* ptrColorIJ, int i, int j, float t)
+	void colorIJ(uchar4* ptrColorIJ, float i, float j, float t)
 	    {
 	    uchar levelGris;
 
@@ -55,20 +54,20 @@ class RipplingMath
 
     private:
 	__device__
-	void f(int i, int j, float t, uchar* ptrlevelGris)
+	void f(float i, float j, float t, uchar* ptrlevelGris)
 	    {
 	    float d;
-	    dij(i,j,&d);
+	    dij(i, j, &d);
 	    *ptrlevelGris = 128.f + 127.f * cosf((d / 10.f) - t / 7.f) / ((d / 10.f) + 1.f);
 
 	    }
 	__device__
-	void dij(int i, int j, float* ptrResult)
+	void dij(float i, float j, float* ptrResult)
 	    {
 	    *ptrResult = sqrtf(f(i) * f(i) + f(j) * f(j));
 	    }
 	__device__
-	float f(int i)
+	float f(float i)
 	    {
 	    return i - dim2;
 	    }
