+++ Student_OMP_Image/src/core/01_Rippling/a_animable/math/RipplingMath.h
@@ -40,7 +40,7 @@ class RipplingMath
 
     public:
 
-	void colorIJ(uchar4* ptrColorIJ, int i, int j, float t)
+	void colorIJ(uchar4* ptrColorIJ, float i, float j, float t)
 	    {
 	    uchar levelGris;
 
@@ -55,22 +55,24 @@ class RipplingMath
 
     private:
 
-	void f(int i, int j, float t, uchar* ptrlevelGris)
+	void f(float i, float j, float t, uchar* ptrlevelGris)
 	    {
 	    float d;
-	    dij(i,j,&d);
+	    dij(i, j, &d);
 	    *ptrlevelGris = 128.f + 127.f * cosf((d / 10.f) - t / 7.f) / ((d / 10.f) + 1.f);
 
 	    }
 
-	void dij(int i, int j, float* ptrResult)
+	void dij(float i, float j, float* ptrResult)
 	    {
 	    *ptrResult = sqrtf(f(i) * f(i) + f(j) * f(j));
 	    }
-	float f(int i)
+
+	float f(float i)
 	    {
 	    return i - dim2;
 	    }
+
 
 	/*--------------------------------------*\
 	|*		Attribut		*|
